Refactored files and classes.
