+++ src/main/java/hhYandexTest/IndexerErrorException.java
@@ -1,0 +1,9 @@
+package hhYandexTest;
+
+public class IndexerErrorException extends Exception {
+    String reason;
+
+    public IndexerErrorException(String reason) {
+        super(reason);
+    }
+}
