+++ src/main/java/hhYandexTest/InverseIndexer/TermManager.java
@@ -1,115 +1,13 @@
-package hhYandexTest;
+package hhYandexTest.InverseIndexer;
 
-import java.io.*;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
-import java.util.*;
-import java.util.function.BiConsumer;
-import java.util.stream.Stream;
+import java.util.Arrays;
+import java.util.Map;
+import java.util.Set;
+import java.util.TreeMap;
 
-public class InverseIndex implements BiConsumer<Long, String> {
-    final static String INVERSE_INDEX_FILE_NAME = "inverse.bin";
-
-    Map<String, TermManager> termMap;
-    String indexDir;
-
-    public InverseIndex(String indexDir) {
-        termMap = new HashMap<>();
-        this.indexDir = indexDir;
-    }
-
-    public void dump() {
-        termMap.forEach((term, termManager) -> {
-           System.out.print(term + ": ");
-           termManager.dump();
-        });
-    }
-
-
-    /*
-        See readme.txt for file structure
-     */
-    public void save() throws IOException {
-        RandomAccessFile    f;
-        Set<String>     terms;
-
-        f = new RandomAccessFile(new File(indexDir, INVERSE_INDEX_FILE_NAME),"rw");
-        terms = termMap.keySet();
-
-        for (String term : terms) {
-            byte [] uIntBytes = {0, 0, 0, 0};
-            byte [] termNameBytes = term.getBytes();
-            long    start = f.getFilePointer();
-            long    end;
-            TermManager termManager = termMap.get(term);
-
-
-            f.write(uIntBytes, 0, 4);  // Will update record size later
-
-            ByteBuffer.allocate(4)
-                    .order(ByteOrder.BIG_ENDIAN)
-                    .putInt(termNameBytes.length)
-                    .rewind()
-                    .get(uIntBytes);
-            f.write(uIntBytes, 0, 4);
-            f.write(termNameBytes, 0, termNameBytes.length);
-
-            byte [] termRawData = termManager.getRawData();
-            f.write(termRawData, 0, termRawData.length);
-
-            end = f.getFilePointer();
-            f.seek(start);
-
-            ByteBuffer.allocate(4)
-                    .order(ByteOrder.BIG_ENDIAN)
-                    .putInt((int) (end - start))
-                    .rewind()
-                    .get(uIntBytes);
-            f.write(uIntBytes, 0, 4);  // Will update record size later
-            f.seek(end);
-        }
-
-        f.close();
-    }
-
-
-    public void addEntry(String term, byte[] document, byte [] metadata) {
-        TermManager termManager;
-
-        if (termMap.containsKey(term) == false) {
-            termManager = new TermManager();
-            termMap.put(term, termManager);
-        } else
-            termManager = termMap.get(term);
-
-        termManager.add(document, metadata);
-    }
-
-
-    public void accept(Long documentNumber, String document) {
-        String doc;
-        byte [] docId = VarByte.longToVarByte(documentNumber);
-
-        // Cleanup
-        doc = document.replaceAll("[!~?#$%^&*_\\-=+'\"><()\\[\\]{}.,:;]", "");
-        doc = doc.replaceAll("\\s+", " ");
-        doc = doc.replaceAll("^\\s+|\\s+$", "");
-        doc = doc.toLowerCase();
-
-        String [] termList  = doc.split(" ");
-
-        for (int n=0;n< termList.length;n++) {
-            if (termList[n].length() < 3)
-                return;
-            addEntry(termList[n], docId, new byte [] {(byte)n});    /* Metadata now is just position */
-        }
-    }
-
-}
-
-
-
-class TermManager {
+public class TermManager {
     final int   JUMP_STEP       = 100;
     final int   DOC_LIST_GROW   = 50;
 
@@ -117,6 +15,9 @@ class TermManager {
     byte [] docList;                        // TODO: Array size is less than 2**32-1, so
     int     docListSize;                    // TODO: may require to switch to ArrayList
 
+    /*
+        Default constructor
+     */
     public TermManager()  {
         jumpTable = new TreeMap<>();
         docList = new byte[DOC_LIST_GROW];
@@ -125,30 +26,66 @@ class TermManager {
 
 
     /*
+        Term constructor with initial data previously obtained with getRawData
+     */
+    public TermManager(byte [] rawTermData) {
+        int offset   = 0;
+        int numSegments = 0;
+
+        for (int n=0;n<4;n++, offset++) {
+            numSegments <<= 8;
+            numSegments |= rawTermData[offset];
+        }
+
+        jumpTable = new TreeMap<>();
+
+        for (int segIdx = 0; segIdx<numSegments; segIdx++) {
+            long    segment     = 0;
+            int     jumpOffset  = 0;
+
+            for (int n=0;n<8; n++, offset++) {
+                segment <<= 8;
+                segment |= rawTermData[offset];
+            }
+
+            for (int n=0;n<4; n++, offset++) {
+                jumpOffset <<= 8;
+                jumpOffset |= rawTermData[offset];
+            }
+
+            jumpTable.put(segment, jumpOffset);
+        }
+
+        docList = Arrays.copyOfRange(rawTermData, offset, rawTermData.length);
+        docListSize = docList.length;
+    }
+
+
+    /*
         TODO: It's better to write directly to file
      */
     public byte [] getRawData() {
 
-        Set<Long>   segments    = jumpTable.keySet();
+        Set<Long> segments    = jumpTable.keySet();
         int jumpTableSize       = 4 + (8 + 4) * segments.toArray().length;
         int resultLength        = jumpTableSize + docListSize;
         byte [] result          = new byte[resultLength];
 
         ByteBuffer resultBuffer = ByteBuffer
-            .allocate(resultLength)
-            .order(ByteOrder.BIG_ENDIAN)
-            .putInt(jumpTableSize);
+                .allocate(resultLength)
+                .order(ByteOrder.BIG_ENDIAN)
+                .putInt(segments.toArray().length);
 
         for (Long segment : segments) {
             resultBuffer
-                .putLong(segment)
-                .putInt(jumpTable.get(segment));
+                    .putLong(segment)
+                    .putInt(jumpTable.get(segment));
         }
 
         resultBuffer
-            .put(docList, 0, docListSize)
-            .rewind()
-            .get(result);
+                .put(docList, 0, docListSize)
+                .rewind()
+                .get(result);
 
         return result;
     }
@@ -217,7 +154,7 @@ class TermManager {
         // Now walk through segment while docIds there <= docId
         //System.out.println("Inserting " + docId + " in segment " + docSegment + " starting from " + position);
 
-    position_seek:
+        position_seek:
         while (position < docListSize) {
             long doc = 0;
             startPos = position;
@@ -266,7 +203,7 @@ class TermManager {
 
             System.arraycopy(docList, 0, newList, 0, position);
             System.arraycopy(
-                docList, position, newList, position + growSize, docListSize - position
+                    docList, position, newList, position + growSize, docListSize - position
             );
 
             // Copy document and metadata
@@ -280,7 +217,7 @@ class TermManager {
         } else {
             //System.out.println(" * Shifting existing array on " + growSize);
             System.arraycopy(
-                docList, position, docList, position + growSize, docListSize - position
+                    docList, position, docList, position + growSize, docListSize - position
             );
 
             System.arraycopy(document, 0, docList, position, document.length);
